requirements.txt
openai
python-dotenv
chromadb
pydantic
numpy
pandas
tqdm
